requirements.txt
pygame-1.9.6